requirements.txt
flask==2.3.2
psycopg2-binary==2.9.6
chromadb==0.3.21
google-generativeai==0.3.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
gunicorn==21.2.0
